Storing one vision of how chunks work that I never committed before I raze everything and start over again.
