+++ gestalt/Cargo.toml
@@ -8,31 +8,33 @@ edition = "2018"
 hemlock = { path = "../hemlock" }
 
 arr_macro = "*"
+base16 = "0.2.1"
 bincode = "1.3.1"
 blake3 = "0.3.7"
 byteorder = "1.3.4"
 cgmath = "0.17.0"
 clap = "2.33.3"
-crossbeam-channel = "0.3"
-custom_error = "1.7.1"
-enum_dispatch = "0.3.2"
+crossbeam-channel = "0.5.0"
+custom_error = "1.8.0"
+enum_dispatch = "0.3.3"
 glium = "0.28.0"
 glutin = { version = "0.25.1", features = ["serde"] }
 hashbrown = "0.9.1"
 image = "0.23.8"
-laminar = "0.4.0"
 lazy_static = "1.4.0"
 legion = "0.3.0"
 log = { version = "0.4.11", features = ["std"] }
 num = "0.3.0"
 parking_lot = "0.11.0"
 rand = "0.7"
-ron = "0.6.1"
+rcgen = "0.8.5"
+ron = "0.6.2"
 rusty_v8 = "0.12.0"
-semver = "0.10.0"
+semver = "0.11.0"
 serde = { version = "1.0.115", features = ["derive"] }
 snow = "0.7.2"
-#sodiumoxide = "0.2"
+sodiumoxide = "0.2.6"
+tokio = "0.3.1"
 ustr = { version = "0.7.0", features = ["serialization"] }
 uuid = { version = "0.8.1", features = ["serde", "v4"] }
