--- gestalt/src/common/identity.rs
@@ -1,3 +1,0 @@
-//TODO: This. 
-
-pub type Identity = u64;
