+++ gestalt/src/common/localization.rs
@@ -1,4 +1,3 @@
 use ustr::*;
 
 pub type LanguageId = Ustr;
-
