+++ gestalt/src/common/mod.rs
@@ -1,4 +1,4 @@
 #[macro_use] pub mod voxelmath;
 #[macro_use] pub mod message;
 pub mod resource;
-pub mod identity;
+#[macro_use] pub mod network;
